+++ tgbotapi.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/send/media/SendMediaGroup.kt
@@ -7,7 +7,11 @@ import dev.inmo.tgbotapi.types.InputMedia.*
 import dev.inmo.tgbotapi.types.MessageIdentifier
 import dev.inmo.tgbotapi.types.chat.abstracts.Chat
 import dev.inmo.tgbotapi.types.message.content.abstracts.MediaGroupContent
+import dev.inmo.tgbotapi.types.message.content.abstracts.VisualMediaGroupContent
+import dev.inmo.tgbotapi.types.message.content.media.AudioContent
+import dev.inmo.tgbotapi.types.message.content.media.DocumentContent
 import dev.inmo.tgbotapi.utils.RiskFeature
+import kotlin.jvm.JvmName
 
 /**
  * @see SendMediaGroup
@@ -33,6 +37,38 @@ suspend fun TelegramBot.sendMediaGroup(
 suspend fun TelegramBot.sendMediaGroup(
     chat: Chat,
     media: List<MediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendMediaGroup(
+    chat.id, media, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
+ * @see SendMediaGroup
+ */
+@RiskFeature(rawSendingMediaGroupsWarning)
+@JvmName("sendMedaGroupByContent")
+suspend fun TelegramBot.sendMediaGroup(
+    chatId: ChatIdentifier,
+    media: List<MediaGroupContent>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendMediaGroup(
+    chatId, media.map { it.toMediaGroupMemberInputMedia() }, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
+ * @see SendMediaGroup
+ */
+@RiskFeature(rawSendingMediaGroupsWarning)
+@JvmName("sendMedaGroupByContent")
+suspend fun TelegramBot.sendMediaGroup(
+    chat: Chat,
+    media: List<MediaGroupContent>,
     disableNotification: Boolean = false,
     protectContent: Boolean = false,
     replyToMessageId: MessageIdentifier? = null,
@@ -72,6 +108,36 @@ suspend fun TelegramBot.sendPlaylist(
 )
 
 /**
+ * @see SendPlaylist
+ */
+@JvmName("sendPlaylistByContent")
+suspend fun TelegramBot.sendPlaylist(
+    chatId: ChatIdentifier,
+    media: List<AudioContent>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendPlaylist(
+    chatId, media.map { it.toMediaGroupMemberInputMedia() }, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
+ * @see SendPlaylist
+ */
+@JvmName("sendPlaylistByContent")
+suspend fun TelegramBot.sendPlaylist(
+    chat: Chat,
+    media: List<AudioContent>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendPlaylist(
+    chat.id, media, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
  * @see SendDocumentsGroup
  */
 suspend fun TelegramBot.sendDocumentsGroup(
@@ -102,6 +168,36 @@ suspend fun TelegramBot.sendDocumentsGroup(
 )
 
 /**
+ * @see SendDocumentsGroup
+ */
+@JvmName("sendDocumentsByContent")
+suspend fun TelegramBot.sendDocumentsGroup(
+    chatId: ChatIdentifier,
+    media: List<DocumentContent>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendDocumentsGroup(
+    chatId, media.map { it.toMediaGroupMemberInputMedia() }, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
+ * @see SendDocumentsGroup
+ */
+@JvmName("sendDocumentsByContent")
+suspend fun TelegramBot.sendDocumentsGroup(
+    chat: Chat,
+    media: List<DocumentContent>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendDocumentsGroup(
+    chat.id, media, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
  * @see SendVisualMediaGroup
  */
 suspend fun TelegramBot.sendVisualMediaGroup(
@@ -123,6 +219,36 @@ suspend fun TelegramBot.sendVisualMediaGroup(
 suspend fun TelegramBot.sendVisualMediaGroup(
     chat: Chat,
     media: List<VisualMediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendVisualMediaGroup(
+    chat.id, media, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
+ * @see SendVisualMediaGroup
+ */
+@JvmName("sendVisualMediaGroupByContent")
+suspend fun TelegramBot.sendVisualMediaGroup(
+    chatId: ChatIdentifier,
+    media: List<VisualMediaGroupContent>,
+    disableNotification: Boolean = false,
+    protectContent: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    allowSendingWithoutReply: Boolean? = null
+) = sendVisualMediaGroup(
+    chatId, media.map { it.toMediaGroupMemberInputMedia() }, disableNotification, protectContent, replyToMessageId, allowSendingWithoutReply
+)
+
+/**
+ * @see SendVisualMediaGroup
+ */
+@JvmName("sendVisualMediaGroupByContent")
+suspend fun TelegramBot.sendVisualMediaGroup(
+    chat: Chat,
+    media: List<VisualMediaGroupContent>,
     disableNotification: Boolean = false,
     protectContent: Boolean = false,
     replyToMessageId: MessageIdentifier? = null,